agent_server/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
langchain
langchain-openai
langchain-mistralai
langchain-google-genai
pydantic
